pyproject.toml
[project]
name = "aegisav"
version = "0.1.0"
description = "Agentic Supervisory Autonomy for Long-Duration Aerial Monitoring"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "AegisAV Team", email = "team@aegisav.org" }
]
keywords = ["drone", "autonomy", "mavlink", "ardupilot", "agentic", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core dependencies
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "pydantic-ai",

    # Web framework
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",

    # MAVLink and vehicle communication
    "pymavlink>=2.4.40",

    # Data and configuration
    "pyyaml>=6.0",
    "numpy>=1.26.0",

    # Logging and monitoring
    "structlog>=24.4.0",
    "logfire[fastapi]>=0.51.0",

    # Persistence
    "redis>=5.0.0",

    # Vision / ML
    "pillow>=10.0.0",
    "ultralytics>=8.2.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",

    # Code quality
    "ruff>=0.5.0",
    "pylint>=3.2.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",

    # Type stubs
    "types-PyYAML>=6.0.12",
    "types-requests>=2.32.0",
]

sim = [
    # Simulation and visualization
    "matplotlib>=3.9.0",
    "plotly>=5.22.0",
    "jupyter>=1.0.0",
    "airsim>=1.8.1",
    "msgpack-rpc-python>=0.4.1",
    "backports.ssl_match_hostname",
]

docs = [
    # Documentation
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]

all = [
    "aegisav[dev,sim,docs]",
]

[project.scripts]
aegis-server = "agent.server.main:main"
aegis-client = "agent.client.main:main"
aegis-demo = "scripts.run_agent:main"

[project.urls]
Homepage = "https://github.com/aegisav/aegisav"
Documentation = "https://aegisav.readthedocs.io/"
Repository = "https://github.com/aegisav/aegisav"
Issues = "https://github.com/aegisav/aegisav/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "autonomy", "metrics", "vision", "simulation"]

# AirSim's build step imports numpy but does not declare it.
[tool.uv.extra-build-dependencies]
airsim = [
    "numpy",
    "msgpack-rpc-python",
    "backports.ssl_match_hostname",
]

# Ruff configuration with AegisAV-specific rules
[tool.ruff]
target-version = "py310"
line-length = 100
show-fixes = true
preview = true

[tool.ruff.lint]
select = [
    # Core Python rules
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade

    # Security and safety rules for autonomous systems
    "S",      # flake8-bandit - security issues
    "A",      # flake8-builtins - builtin function usage

    # Type annotations - enforce types everywhere
    "ANN",    # flake8-annotations - type hints for functions/methods

    # Docstrings - enforce documentation
    "D",      # pydocstyle - docstring rules

    # Code quality rules
    "ARG001", # unused arguments in functions
    "ARG002", # unused method arguments
    "RUF",    # ruff-specific rules
    "SIM108", # ternary operator
    "T",      # flake8-print - print statements
    "PLC",    # pylint conventions
]

ignore = [
    "E501",   # line too long, handled by black
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex (handled by complexity check)
    "S101",   # assert usage (allowed in tests)
    "S311",   # random usage (allowed for simulation)

    # Type annotation exceptions
    "ANN101", # missing type for self (redundant)
    "ANN102", # missing type for cls (redundant)
    "ANN401", # allow Any type when necessary

    # Docstring exceptions
    "D100",   # missing docstring in public module (module docstrings optional)
    "D104",   # missing docstring in public package
    "D105",   # missing docstring in magic method
    "D107",   # missing docstring in __init__
    "D203",   # no blank line before class docstring (conflicts with D211)
    "D213",   # multi-line summary second line (conflicts with D212)
    "D415",   # first line should end with punctuation (too pedantic for short docstrings)
]

# AegisAV-specific rule configurations
[tool.ruff.lint.per-file-ignores]
"agent/server/decision.py" = ["RUF009"]  # Allow complex expressions in decision logic
"agent/server/world_model.py" = ["RUF005"]  # Allow complex loops in world modeling
"autonomy/mavlink_interface.py" = ["RUF012"]  # Allow try/except in MAVLink comms
"agent/server/critics/*.py" = ["PLC0415"]  # Allow in-function pydantic_ai imports (optional dep)
"tests/*" = ["S", "PLC0415", "ANN", "D"]  # Tests: security, imports, annotations, docstrings
"scripts/*" = ["D"]  # Scripts: docstrings optional
"examples/*" = ["D", "ANN"]  # Examples: relaxed rules

# Pydocstyle configuration - Google-style docstrings
[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.ruff.lint.flake8-copyright]
# Copyright notice requirements
notice-rgx = "Copyright \\d{4,} AegisAV Team"
min-file-size = 1024

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["agent", "autonomy", "metrics"]

# Pylint configuration
[tool.pylint.main]
py-version = "3.10"

[tool.pylint.format]
max-line-length = 100

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
    "duplicate-code",
    "fixme",
    "import-error",
    "no-name-in-module",
    "logging-fstring-interpolation",
    "broad-exception-caught",
]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pymavlink.*",
    "structlog.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=agent",
    "--cov=autonomy",
    "--cov=metrics",
    "--cov=vision",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "mavlink: marks tests that require MAVLink connection",
]

# Coverage configuration
[tool.coverage.run]
source = ["agent", "autonomy", "metrics", "vision"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
